# api/openapi.yaml
openapi: 3.0.3
info:
  title: My Hike API
  version: 1.0.0
  description: |
    REST API for My Hike. Except for registration, login, health, and this
    document, endpoints require a JWT access token in the Authorization header.
    Administrator endpoints additionally require the `admin` role.
servers:
  - url: http://localhost:3000
    description: Local development server
...
paths:
...
  /api/auth/register:
    post:
      tags: [Authentication]
      summary: Register a user
      operationId: registerUser
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterRequest"
      responses:
        "201":
          description: User registered
          content:
            application/json:
              schema:
                type: object
                required: [user]
                properties:
                  user:
                    $ref: "#/components/schemas/PublicUser"
        "400":
          $ref: "#/components/responses/BadRequest"
        "409":
          description: Email is already registered
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /api/auth/login:
    post:
      tags: [Authentication]
      summary: Log in and obtain an access token
      operationId: loginUser
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginRequest"
      responses:
        "200":
          description: Login successful
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
...
  /api/friends:
    get:
      tags: [Friends]
      summary: List accepted friends
      operationId: listFriends
      responses:
        "200":
          description: Accepted friends
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/PublicUser"
        "401":
          $ref: "#/components/responses/Unauthorized"
    post:
      tags: [Friends]
      summary: Send a friend request by exact email address
      operationId: createFriendRequest
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email]
              properties:
                email:
                  type: string
                  format: email
                  example: friend@example.com
      responses:
        "201":
          description: Friend request created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FriendRequest"
        "400":
          $ref: "#/components/responses/BadRequest"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          description: Connection or request already exists
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /api/friends/search:
    get:
      tags: [Friends]
      summary: Search active users by partial name or exact email
      operationId: searchFriends
      parameters:
        - name: search
          in: query
          required: true
          schema:
            type: string
            maxLength: 320
      responses:
        "200":
          description: Up to 20 matching users
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/FriendSearchResult"
        "401":
          $ref: "#/components/responses/Unauthorized"
  /api/friends/requests:
    get:
      tags: [Friends]
      summary: List incoming and outgoing pending friend requests
      operationId: listFriendRequests
      responses:
        "200":
          description: Pending requests
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/FriendRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
  /api/friends/requests/{requestId}:
    delete:
      tags: [Friends]
      summary: Reject or cancel a pending friend request
      operationId: deleteFriendRequest
      parameters:
        - $ref: "#/components/parameters/RequestIdPath"
      responses:
        "204":
          description: Request removed
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
...
  /api/friends/{friendId}:
    delete:
      tags: [Friends]
      summary: Remove an accepted friend
      operationId: deleteFriend
      parameters:
        - name: friendId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "204":
          description: Friend removed
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
  /api/friends/comparison:
    get:
      tags: [Friends]
      summary: Compare daily activity counts with accepted friends
      operationId: compareFriendActivities
      parameters:
        - name: friendIds
          in: query
          description: Comma-separated friend UUIDs; maximum 50
          schema:
            type: string
        - $ref: "#/components/parameters/MonthQuery"
      responses:
        "200":
          description: Current user and selected friend series
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/FriendComparisonSeries"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
...
  /api/friends/activities/{activityId}/reactions:
    post:
      tags: [Friend activities]
      summary: Add a like or slap to an accepted friend's activity
      operationId: createActivityReaction
      parameters:
        - $ref: "#/components/parameters/ActivityIdPath"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ReactionInput"
      responses:
        "200":
          description: Reaction already existed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ActivityReaction"
        "201":
          description: Reaction created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ActivityReaction"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
  /api/friends/activities/{activityId}/reactions/{type}:
    delete:
      tags: [Friend activities]
      summary: Remove the user's reaction from a friend's activity
      operationId: deleteActivityReaction
      parameters:
        - $ref: "#/components/parameters/ActivityIdPath"
        - $ref: "#/components/parameters/ReactionTypePath"
      responses:
        "204":
          description: Reaction removed
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
...
    RequestIdPath:
      name: requestId
      in: path
      required: true
      schema:
        type: string
        format: uuid
    ActivityIdPath:
      name: activityId
      in: path
      required: true
      schema:
        type: string
        format: uuid
    ReactionTypePath:
      name: type
      in: path
      required: true
      schema:
        $ref: "#/components/schemas/ReactionType"
...
    MonthQuery:
      name: month
      in: query
      schema:
        type: string
        pattern: '^\d{4}-(0[1-9]|1[0-2])$'
        example: "2026-09"
...
  responses:
    BadRequest:
      description: Invalid input
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    Unauthorized:
      description: Missing, invalid, expired, or deleted-user token
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    Forbidden:
      description: The authenticated user is not allowed to perform this action
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    NotFound:
      description: Resource not found or not owned by the authenticated user
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
...
  schemas:
    Error:
      type: object
      required: [error]
      properties:
        error:
          type: string
...
    UserRole:
      type: string
      enum: [normal_user, admin]
...
    PublicUser:
      type: object
      required: [id, name, email, role]
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        email:
          type: string
          format: email
        role:
          $ref: "#/components/schemas/UserRole"
...
    RegisterRequest:
      type: object
      required: [name, email, password, repeatPassword]
      properties:
        name:
          type: string
        email:
          type: string
          format: email
        password:
          type: string
          format: password
          minLength: 8
        repeatPassword:
          type: string
          format: password
          minLength: 8
    LoginRequest:
      type: object
      required: [email, password]
      properties:
        email:
          type: string
          format: email
        password:
          type: string
          format: password
          minLength: 8
...
    AuthResponse:
      type: object
      required: [token, user]
      properties:
        token:
          type: string
          description: JWT access token valid for seven days
        user:
          $ref: "#/components/schemas/PublicUser"
...
    ReactionType:
      type: string
      enum: [like, slap]
    ReactionInput:
      type: object
      required: [type]
      properties:
        type:
          $ref: "#/components/schemas/ReactionType"
    ActivityReaction:
      type: object
      required: [activityId, activityOwnerId, userId, type, createdAt]
      properties:
        activityId:
          type: string
          format: uuid
        activityOwnerId:
          type: string
          format: uuid
        userId:
          type: string
          format: uuid
        type:
          $ref: "#/components/schemas/ReactionType"
        createdAt:
          type: string
          format: date-time
    FriendSearchResult:
      allOf:
        - $ref: "#/components/schemas/PublicUser"
        - type: object
          required: [connectionStatus]
          properties:
            connectionStatus:
              type: string
              enum: [none, pending, accepted]
    FriendRequest:
      type: object
      required: [id, direction, user, createdAt]
      properties:
        id:
          type: string
          format: uuid
        direction:
          type: string
          enum: [incoming, outgoing]
        user:
          $ref: "#/components/schemas/PublicUser"
        createdAt:
          type: string
          format: date-time
...
    FriendComparisonDay:
      type: object
      required: [date, activityCount]
      properties:
        date:
          type: string
          format: date
        activityCount:
          type: integer
          minimum: 0
    FriendComparisonSeries:
      type: object
      required: [userId, name, isCurrentUser, days]
      properties:
        userId:
          type: string
          format: uuid
        name:
          type: string
        isCurrentUser:
          type: boolean
        days:
          type: array
          items:
            $ref: "#/components/schemas/FriendComparisonDay"
...
security:
  - bearerAuth: []